Add common functionality in utils
